+++ src/main/java/com/onevizion/uitest/api/helper/entity/EntityConfigFieldEfileMetadata.java
@@ -57,7 +57,7 @@ public class EntityConfigFieldEfileMetadata {
 
         new Select(seleniumSettings.getWebDriver().findElement(By.name(TYPE))).selectByVisibleText(configFieldVoEfileMetadata.getType().getLabel());
 
-        selector.selectRadio(By.name(FIELD_BUTTON), By.id(AbstractSeleniumCore.BUTTON_OK_ID_BASE + 0L), 1L, configFieldVoEfileMetadata.getConfigField(), 1L);
+        selector.selectRadio(By.name(FIELD_BUTTON), By.id(AbstractSeleniumCore.BUTTON_OK_ID_BASE + 0L), 1, configFieldVoEfileMetadata.getConfigField(), 1L);
 
         window.closeModal(By.id(AbstractSeleniumCore.BUTTON_OK_ID_BASE));
         grid2.waitLoad(3L);
@@ -75,7 +75,7 @@ public class EntityConfigFieldEfileMetadata {
     }
 
     public void testInGrid(Long gridId, Long rowIndex, ConfigFieldVoEfileMetadata configFieldVoEfileMetadata) {
-        Map<Long, String> gridVals = new HashMap<>();
+        Map<Integer, String> gridVals = new HashMap<>();
 
         gridVals.put(js.getColumnIndexByLabel(gridId, "Metadata Type"), configFieldVoEfileMetadata.getType().getLabel());
         gridVals.put(js.getColumnIndexByLabel(gridId, "Metadata Field"), configFieldVoEfileMetadata.getConfigField());
